Crate move_unit_test

Modules

cargo_runner

extensions

This module manages native extensions supported by the unit testing framework. Such extensions are enabled by cfg features and must be compiled into the test to be usable.

test_reporter

test_runner

Macros

register_move_unit_tests

Structs

UnitTestingConfig